requirements.txt
Axelrod==1.6.0
cycler==0.10.0
dj-database-url==0.4.1
Django==1.10
django-debug-toolbar==1.5
djangorestframework==3.4.1
gunicorn==19.6.0
hypothesis==3.4.2
matplotlib==1.5.1
numpy==1.11.1
psycopg2==2.6.2
pyparsing==2.1.1
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
sqlparse==0.2.1
tqdm==3.4.0
whitenoise==3.2
